refactor: TableGroup 관련 도메인 리팩터링 - 의존성 개선에 따라 의존성 변경 - 메서드 변경
test: 로직 변경에 따라 테스트 코드 수정
